requirements-py312.txt
# Python 3.12 Compatible Requirements
# Updated versions for Python 3.12 compatibility

# System Dependencies
# Poppler is required for PDF processing. Install via:
# Windows: winget install poppler
# Linux: sudo apt-get install poppler-utils
# macOS: brew install poppler

# Core Framework
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
waitress==2.1.2

# Security
bcrypt==4.1.2
cryptography==44.0.1

# Document Processing
# Note: Using compatible versions for Python 3.12
unstructured==0.10.30
unstructured-inference==0.7.21
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10

# Vector Database
supabase==2.8.1

# ML & Embeddings - Python 3.12 compatible versions
torch==2.2.0
transformers==4.35.2
sentence-transformers==2.7.0
huggingface-hub==0.19.4
langchain-community==0.0.19

# Utilities - Python 3.12 compatible versions
pillow==10.2.0
tqdm==4.66.1
nltk==3.8.1
requests==2.31.0
numpy>=1.26.0  # Important: 1.26+ for Python 3.12
anyio<4.0

# Production Server
gunicorn==21.2.0

# Build tools
setuptools>=68.0.0
wheel>=0.40.0